Extracting the styles
