+++ events/features/indexer-widget/page-schedule-indexer.css
@@ -1,0 +1,115 @@
+.page-path-indexer-widget {
+  position: fixed;
+  display: flex;
+  flex-direction: column;
+  bottom: 12px;
+  right: 0;
+  z-index: 1000;
+  box-shadow: 2px 2px 10px 0 rgb(0 0 0 / 50%);
+  border-radius: 16px 0 0 16px;
+  background-color: var(--color-gray-700);
+  color: var(--color-white);
+  padding: 12px 12px 32px;
+  font-family: var(--font-family-body);
+  font-size: var(--type-body-xxs-size);
+  transition: transform 0.3s ease;
+  transform: translateX(calc(100% - 12px));
+
+  .page-path-indexer-hover-tab {
+    position: absolute;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: 40px;
+    top: 50%;
+    left: -16px;
+    transform: translateY(-50%);
+    background-color: var(--color-gray-700);
+    color: var(--color-white);
+    padding: 8px 4px;
+    border-radius: 8px 0 0 8px;
+    font-size: var(--type-body-xxs-size);
+    font-weight: 700;
+    text-orientation: mixed;
+  }
+
+  /* Visible state - slides fully into view */
+  &:not(.hidden) {
+    transform: translateX(0);
+  }
+
+  .page-path-indexer-schedule-id-list {
+    padding: 16px;
+    font-size: var(--type-body-xs-size);
+    border-bottom: 1px solid var(--color-gray-600);
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+
+    .page-path-indexer-schedule-id-list-title {
+      font-weight: 700;
+      margin-bottom: 8px;
+      font-size: var(--type-body-m-size);
+    }
+
+    .page-path-indexer-schedule-id-item {
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
+      gap: 8px;
+      height: 32px;
+      white-space: nowrap;
+    }
+
+    .page-path-indexer-schedule-id-item-status-wrapper {
+      align-items: center;
+      gap: 4px;
+      display: none;
+    }
+
+    .page-path-indexer-schedule-id-item-status-text {
+      font-size: var(--type-body-xs-size);
+      font-family: var(--font-family-body);
+      font-weight: 700;
+    }
+
+    button.page-path-indexer-schedule-id-item-status-text {
+      cursor: pointer;
+    }
+
+    button.page-path-indexer-schedule-id-item-status-text:disabled {
+      cursor: not-allowed;
+      opacity: 0.5;
+      pointer-events: none;
+    }
+
+    .page-path-indexer-schedule-id-item-status-wrapper.indexed {
+      color: var(--color-green);
+    }
+
+    .page-path-indexer-schedule-id-item-status-wrapper.not-indexed {
+      color: var(--color-red);
+    }
+
+    .page-path-indexer-schedule-id-item.indexed .page-path-indexer-schedule-id-item-status-wrapper.indexed {
+      display: flex;
+    }
+
+    .page-path-indexer-schedule-id-item:not(.indexed) .page-path-indexer-schedule-id-item-status-wrapper.not-indexed {
+      display: flex;
+    }
+  }
+    
+  button.page-path-indexer-button {
+    margin: 16px;
+    cursor: pointer;
+    font-family: var(--font-family-body);
+    font-size: var(--type-body-xs-size);
+  }
+
+  button.page-path-indexer-button:disabled {
+    cursor: not-allowed;
+    opacity: 0.5;
+    pointer-events: none;
+  }
+}
